Make dedicated snapshot test classes
